+++ test-fixtures/string-esm-config/test-prettier-config-esm/index.js
@@ -1,0 +1,9 @@
+// ESM Prettier config package
+const config = {
+  semi: true,
+  singleQuote: false,
+  tabWidth: 2,
+  trailingComma: "es5",
+};
+
+export default config;